First iteration on KV service to make it extensible.
