+++ core/src/main/resources/application.conf
@@ -283,7 +283,7 @@ backend {
 
 services {
   KeyValue {
-    class = "cromwell.services.keyvalue.impl.KeyValueServiceActor"
+    class = "cromwell.services.keyvalue.impl.SqlKeyValueServiceActor"
   }
   MetadataService {
     class = "cromwell.services.metadata.impl.MetadataServiceActor"
